Finished base for header / footer and sign data
This is the first real commit as i have finished the header and nav. I
also completely finished the layout of everything on desktop. Now i have
to work on finishing layout of body sections for each page. Next will be
the messaging center as its hardest / most complete idea.