Improved message center and replaced requires
I arguably improved the layout of the messaging center as well as
replaced all of the content that had the require("common.php"); to a
require once to not try to change or replace the header. Next i hope to
work on a better message insert system and make it more fancy.